Updated type for index into errors_get_message_of_code

+++ sources/errors/get.c
@@ -11,7 +11,7 @@
 char *errors_get_message_of_code(code_msg_t *messages, size_t messages_size,
 int code)
 {
-    for (int i = 0; i < (int)messages_size; i++) {
+    for (size_t i = 0; i < messages_size; i++) {
         if (messages[i].code == code)
             return messages[i].msg;
     }
